# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: A Todo List Microservice
  description: Title
  version: 1.0.0
servers:
  - url: http://localhost:8080/
paths:
  /users:
    post:
      tags:
        - createUser
      summary: Create a new user for system
      operationId: createUser
      requestBody:
        description: Create a new user for system
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: Create user successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectCreationSuccessResponse'

    get:
      tags:
        - getUserList
      summary: Get the list of user in this system
      operationId: GetListUser
      responses:
        '200':
          description: Get the list of user in system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'

    put:
      tags:
        - UpdateUser
      summary: Update User Information
      operationId: UpdateUser
      requestBody:
        description: Update User Info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Update Category successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectCreationSuccessResponse'

  /users/{id}:
    get:
      tags:
        - getUserbyId
      summary: Get User by Id
      operationId: GetUserId
      parameters:
        - name: id
          in: path
          description: UserId
          required: true
          schema:
            type: string
            format: string
      responses:
        '200':
          description: Get the info of user by Id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
    delete:
      tags:
        - Delete User Id
      summary: Delete User
      operationId: DeleteUser
      parameters:
        - name: id
          in: path
          description: UserId
          required: true
          schema:
            type: string
            format: string
      responses:
        '200':
          description: Delete User successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectCreationSuccessResponse'
  /todo:
    post:
      tags:
        - createtodo
      summary: Create a todo task
      operationId: createTodo
      requestBody:
        description: Create a Todo task for system
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/CreateToDoRequest'
      responses:
        '200':
          description: Create Todo successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectCreationSuccessResponse'
    get:
      tags:
        - getTodoList
      summary: Get the list of todo in this system
      operationId: GetListTask
      responses:
        '200':
          description: Get the list of task in system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
  /todo/{toDoid}:
    put:
      tags:
        - UpdateTask
      summary: Update Task Information by Id
      operationId: UpdateTask
      parameters:
        - name: toDoid
          in: path
          description: TaskId
          required: true
          schema:
            type: string
            format: string
      requestBody:
        description: Update Task Info by Id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Update Task successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectCreationSuccessResponse'


components:
  schemas:
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponseModel'
    UserResponseModel:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        username:
          type: string
        userId:
          type: string
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        username:
          type: string
        password:
          type: string
    UserInfoResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserInfoResponseModel'
    UserInfoResponseModel:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        username:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        userId:
          type: string
        Username:
          type: string
        Name:
          type: string
        Email:
          type: string
    CreateToDoRequest:
      type: object
      properties:
        authorId:
          type: string
        tittle:
          type: string
        Content:
          type: string
        priority:
          type: string
        status:
          type: boolean
          default: false
    TaskListResponse:
      type: object
      properties:
        todo:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponseModel'
    TaskResponseModel:
      type: object
      properties:
        authorId:
          type: string
        authorName:
          type: string
        tittle:
          type: string
        content:
          type: string
        priority:
          type: string
        status:
          type: boolean
    UpdateTaskRequest:
      type: object
      properties:
        tittle:
          type: string
        content:
          type: string
        priority:
          type: string
        status:
          type: boolean

    ObjectCreationSuccessResponse:
      type: object
      properties:
        id:
          type: string
        responseCode:
          type: integer

    ApiError:
      type: object
      properties:
        faults:
          type: array
          items:
            $ref: "#/components/schemas/Fault"
        responseCode:
          type: integer
    Fault:
      type: object
      properties:
        internalMessage:
          type: string
        code:
          type: integer
        userMessage:
          type: string

  examples:
    InternalServerError:
      value:
        responseCode: 500
        faults:
          -:
          userMessage: There're some error when processing this request. Please retry again or contact the site administrator for supporting
          internalMessage: Internal server error
          code: 1000
          responseCode: 500